Add files via upload
This project aims to find cheap and good restaurants near me.
The first step is to create a database by webscraping Ifood website (the platform which I will order food from).
The webscraping part of the project is done by using the selenium library and respecting the website documentation.
In the future, the recommendation of which item from which restaurant will be made by using machine learning. The algorithm will be fed with the webscraped data and my personal info on previous purchases.